+++ NEWS.rst
@@ -6,7 +6,7 @@ of every change, see the Git log.
 
 Latest
 ------
-* Major: Depend on options registry to access Git functionality during
+* Minor: Depend on options registry to access Git functionality during
   options.
 
 4.25.0